% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_ifzero}
\alias{add_ifzero}
\title{Add a fixed value to zero counts (used for zero counts in Laplacian approximation of binomial likelihood)}
\usage{
add_ifzero(y_ij, value)
}
\arguments{
\item{y_ij}{vector of per-feature read count across samples}
}
\value{
modified counts
}
\description{
Add a fixed value to zero counts (used for zero counts in Laplacian approximation of binomial likelihood)
}
