% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseDOSSA2_fit.R
\name{control_fit}
\alias{control_fit}
\title{Control options for fit_SparseDOSSA2 and fitCV_SparseDOSSA2}
\usage{
control_fit(
  maxit = 100,
  rel_tol = 0.01,
  abs_tol = 0.01,
  control_numint = list(),
  verbose = FALSE,
  debug_dir = NULL
)
}
\arguments{
\item{maxit}{maximum number of EM iterations}

\item{rel_tol}{relative change threshold in the log likelihood
for algorithm convergence}

\item{abs_tol}{absolute change threshold in the log likelihood
for algorithm convergence}

\item{control_numint}{a named list of control parameters for the 
numerical integrations during the E step. See help page for
\code{control_numint}}

\item{verbose}{whether or not detailed running messages should be provided}

\item{debug_dir}{directory for intermediate output, such as the
EM expectations and parameter values and during each step of the
EM algorithm. Default to \code{NULL} in which case no such output
will be generated}
}
\value{
a list of the same names
}
\description{
Control options for fit_SparseDOSSA2 and fitCV_SparseDOSSA2
}
