% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_expectation.R
\name{control_integrate}
\alias{control_integrate}
\title{control parameters for the 
numerical integrations during the E step of SparseDOSSA2's fitting}
\usage{
control_integrate(
  rel_tol = 0.01,
  abs_tol = 0,
  max_eval = 50,
  maxit_limits = 10,
  precBits = 200,
  only_value = TRUE
)
}
\arguments{
\item{rel_tol}{relative change threshold in the integration values for the 
integration to converge}

\item{abs_tol}{absolute change threshold in the integration values for the 
integration to converge}

\item{max_eval}{maximum of integration evaluations allowed}

\item{maxit_limits}{maximum number of tries allowed to guess the integration's
lower and upper limits}

\item{precBits}{numeric precision used for the integration values}

\item{only_value}{whether or not only the integration value should be returned}
}
\value{
a list of the same names
}
\description{
control parameters for the 
numerical integrations during the E step of SparseDOSSA2's fitting
}
