% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor2.R
\name{cor2}
\alias{cor2}
\title{correlation estimation with randomized option to break ties}
\usage{
cor2(x, y = x, method = "spearman", random = FALSE, R = 30, seed)
}
\arguments{
\item{x}{numeric matrix x}

\item{y}{numeric matrix y (set to x if unprovided)}

\item{method}{method used for correlation estimation. Only "spearman" and "kendall" are supported.}

\item{random}{if randomization should be performed in order to break ties}

\item{R}{number of random iterations to perform}

\item{seed}{random seed}
}
\value{
estimated correlation matrix
}
\description{
correlation estimation with randomized option to break ties
}
