% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fromMetaphlan.R
\name{create_template_fromMetaphlan}
\alias{create_template_fromMetaphlan}
\title{Create "count" template from Metaphlan output}
\usage{
create_template_fromMetaphlan(feature_abd, n_i)
}
\arguments{
\item{feature_abd}{Metaphlan feature abundance table}

\item{n_i}{Metaphlan estimate of read depth}
}
\value{
"Count" feature table that can be used as template for SparseDOSSA2
}
\description{
Create "count" template from Metaphlan output
}
