% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_diagnose.R
\name{EM_diagnose}
\alias{EM_diagnose}
\title{Title}
\usage{
EM_diagnose(data, params_init = NULL, control = list())
}
\arguments{
\item{data}{}

\item{params_init}{}

\item{control}{}
}
\value{

}
\description{
Title
}
