% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_C}
\alias{estimate_C}
\title{Normal copula estimator for feature-feature association (correlation between the p_ijs)}
\usage{
estimate_C(feature_abd, control = list(method = "spearman", random =
  TRUE, R = 50))
}
\arguments{
\item{feature_abd}{feature x sample abundance table}

\item{control}{additional control parameters for correlation estimation}
}
\value{
the estimated copula
}
\description{
Normal copula estimator for feature-feature association (correlation between the p_ijs)
}
