% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_F}
\alias{estimate_F}
\title{Non-parametric density estimator for the distribution of per-feature parameters}
\usage{
estimate_F(feature_params, control = list(estimator = "Hscv"))
}
\arguments{
\item{feature_params}{data frame of estimated per-feature parameters (in order mu, sigma2, and pi)}

\item{control}{list of control parameters to pass on to the bandwidth estimators in ks}
}
\value{
list with components for bandwidth estimations for zero and non-zero sigma2 part, and
percentage of zero sigma2s
}
\description{
Non-parametric density estimator for the distribution of per-feature parameters
}
