% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_featureParam}
\alias{estimate_featureParam}
\title{EM algorithm estimator for per-feature parameters (prevalence, mean, variance)}
\usage{
estimate_featureParam(y_ij, n_i, control = list(maxiter.outer = 100,
  maxiter.inner = 1000, reltol.outer = 1e-08, factr.inner = 1e+07))
}
\arguments{
\item{y_ij}{vector of per-feature read count across samples}

\item{n_i}{vector of total read count across samples}

\item{control}{list of control parameters of EM algorithm}
}
\value{
list with components for estimated parameters, hidden parameters, and optimization information
}
\description{
EM algorithm estimator for per-feature parameters (prevalence, mean, variance)
}
