% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_readCount}
\alias{estimate_readCount}
\title{MLE estimator for read depth distribution}
\usage{
estimate_readCount(n_i)
}
\arguments{
\item{n_i}{vector of total read count across samples}
}
\value{
estimated mean and variance parameters
}
\description{
MLE estimator for read depth distribution
}
