% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseDOSSA2_fit.R
\name{fit_SparseDOSSA2}
\alias{fit_SparseDOSSA2}
\title{Title}
\usage{
fit_SparseDOSSA2(data, lambda = NULL, control = list())
}
\arguments{
\item{data}{}

\item{lambda}{}

\item{control}{}
}
\value{

}
\description{
Title
}
