% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseDOSSA2_fit.R
\name{fitCV_SparseDOSSA2}
\alias{fitCV_SparseDOSSA2}
\title{Title}
\usage{
fitCV_SparseDOSSA2(data, lambdas, K = 5, control = list())
}
\arguments{
\item{data}{}

\item{lambdas}{}

\item{K}{}

\item{control}{}
}
\value{

}
\description{
Title
}
