% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_basis}
\alias{generate_basis}
\title{Generate basis (i.e., pij) matrix, with the option to incorporate copula shuffling for correlation structure}
\usage{
generate_basis(feature_param, n_sample, fit_C = NULL, seed)
}
\arguments{
\item{feature_param}{data frame/matrix of generated (if same_features = FALSE) or estimated
(if same_feature = TRUE) per-feature parameters (in order mu, sigma2, and pi)}

\item{n_sample}{number of samples to generate}

\item{fit_C}{a copula fit (fit generated from estimate_C). If null then no shuffling will be performed.}

\item{seed}{random seed}
}
\value{
a list of matrices, in order:
mat_basis_original is the originally generated matrix
mat_basis_shuffled is the copula shuffled matrix (is equal to mat_basis_original if fit_C is null)
mat_basis is the normalized basis matrix for proper multinomial sampling
}
\description{
Generate basis (i.e., pij) matrix, with the option to incorporate copula shuffling for correlation structure
}
