% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_feature_metadata_spike_df}
\alias{generate_feature_metadata_spike_df}
\title{Title}
\usage{
generate_feature_metadata_spike_df(
  features,
  perc_feature_spiked_metadata,
  n_metadata,
  effect_size,
  spike_metadata = c("both", "abundance", "prevalence")
)
}
\arguments{
\item{features}{}

\item{perc_feature_spiked_metadata}{}

\item{n_metadata}{}

\item{effect_size}{}

\item{spike_metadata}{}
}
\value{

}
\description{
Title
}
