% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_featureParam}
\alias{generate_featureParam}
\title{Generate per-feature parameters (mu, sigma, pi) given pre-fit non-parameter density kernels}
\usage{
generate_featureParam(fit_F, feature_param, n_feature)
}
\arguments{
\item{fit_F}{a list (fit generated fro estimate_F)}

\item{feature_param}{data frame of estimated per-feature parameters (in order mu, sigma2, and pi)}

\item{n_feature}{number of features to generate}
}
\value{
matrix of per-feature parameters (in order mu, sigma2, and pi)
}
\description{
Generate per-feature parameters (mu, sigma, pi) given pre-fit non-parameter density kernels
}
