% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_readCount}
\alias{generate_readCount}
\title{Generate multinomial sampled read counts matrix}
\usage{
generate_readCount(mat_p, n_i, seed)
}
\arguments{
\item{mat_p}{feature x sample matrix of underlying relative abundance}

\item{n_i}{vector of total read count across samples}
}
\value{
feature x sample matrix of microbial read count
}
\description{
Generate multinomial sampled read counts matrix
}
