% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_readDepth}
\alias{generate_readDepth}
\title{Generate log-normal per-sample read counts}
\usage{
generate_readDepth(mu, sigma2, n_sample, read_depth)
}
\arguments{
\item{mu}{mean parameter}

\item{sigma2}{variance parameter}

\item{n_sample}{number of samples to generate}

\item{read_depth}{median read depth to normalize to}
}
\value{
vector of total read depths
}
\description{
Generate log-normal per-sample read counts
}
