% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_ZILogitN}
\alias{generate_ZILogitN}
\title{Generate zero-inflated logistic normal per-feature basis (pij for fixed j)}
\usage{
generate_ZILogitN(mu, sigma2, pi, n_sample)
}
\arguments{
\item{mu}{mean parameter for the normal component}

\item{sigma2}{variance parameter for the normal component}

\item{pi}{prevalence (i.e. 1 - zero proportion) parameter}

\item{n_sample}{number of samples to generate}
}
\value{
vector of zero-infated proportions for a given feature
}
\description{
Generate zero-inflated logistic normal per-feature basis (pij for fixed j)
}
