% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_helpers.R
\name{get_sigma2hat_ij}
\alias{get_sigma2hat_ij}
\title{per-feature MLE variance of mu_ij (= logit p_ij) across samples by binomial model}
\usage{
get_sigma2hat_ij(y_ij, n_i)
}
\arguments{
\item{y_ij}{vector of per-feature read count across samples}

\item{n_i}{vector of total read count across samples}
}
\value{
vector sigma2hat_ij
}
\description{
per-feature MLE variance of mu_ij (= logit p_ij) across samples by binomial model
}
