% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_w_i}
\alias{get_w_i}
\title{E step update for the per sample weights of prevalence parameter}
\usage{
get_w_i(muhat_ij, sigma2hat_ij, y_ij, n_i, pi, mu, sigma2)
}
\arguments{
\item{muhat_ij}{vector of per-feature muhat estimation across samples}

\item{sigma2hat_ij}{vector of per-feature sigma2hat estimation across samples}

\item{y_ij}{vector of per-feature read count across samples}

\item{n_i}{vector of total read count across samples}

\item{pi}{pi parameter (in current iteration)}

\item{mu}{mu parameter (in current iteration)}

\item{sigma2}{sigma2 parameter (in current iteration)}
}
\value{
weights of observations, in the form as a) expectation (probability) of sequencing (as
opposed to biological) zero for zero-count samples, and b) ones for non-zero-count samples
}
\description{
E step update for the per sample weights of prevalence parameter
}
