% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{negLogLik_sigma2}
\alias{negLogLik_sigma2}
\title{Negative log likelihood with respect to sigma2}
\usage{
negLogLik_sigma2(sigma2, muhat_ij, sigma2hat_ij, w_i)
}
\arguments{
\item{sigma2}{sigma2 parameter}

\item{muhat_ij}{vector of per-feature muhat estimation across samples}

\item{sigma2hat_ij}{vector of per-feature sigma2hat estimation across samples}

\item{w_i}{vector of observation weights}
}
\value{
negative log likelihood, evaluated at sigma2
}
\description{
Negative log likelihood with respect to sigma2
}
