% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_helpers.R
\name{posterior_mu_ij}
\alias{posterior_mu_ij}
\title{Posterior estimation of per-sample mu (shrinked by the estimated parameters)}
\usage{
posterior_mu_ij(muhat_ij, sigma2hat_ij, mu, sigma2)
}
\arguments{
\item{muhat_ij}{vector of per-feature muhat estimation across samples (before shrinkage)}

\item{sigma2hat_ij}{vector of per-feature sigma2hat estimation across samples (before shrinkage)}

\item{mu}{estimated mu parameter}

\item{sigma2}{estimated sigma2 parameter}
}
\value{
posterior muhat_ij, shrinked towards mu
}
\description{
Posterior estimation of per-sample mu (shrinked by the estimated parameters)
}
