% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{posterior_sigma2_ij}
\alias{posterior_sigma2_ij}
\title{Posterior estimation of per-sample sigma2 (shrinked by the estimated parameters)}
\usage{
posterior_sigma2_ij(sigma2hat_ij, sigma2)
}
\arguments{
\item{sigma2hat_ij}{vector of per-feature sigma2hat estimation across samples (before shrinkage)}

\item{sigma2}{estimated sigma2 parameter}
}
\value{
posterior sigma2hat_ij
}
\description{
Posterior estimation of per-sample sigma2 (shrinked by the estimated parameters)
}
