% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseDOSSA2.R
\name{SparseDOSSA2}
\alias{SparseDOSSA2}
\title{Simulate synthetic microbial abundance observations with SparseDOSSA2}
\usage{
SparseDOSSA2(
  template = "Stool",
  n_sample = 100,
  new_features = TRUE,
  n_feature = 100,
  spike_metadata = "none",
  metadata_effect_size = 1,
  perc_feature_spiked_metadata = 0.05,
  metadata_matrix = NULL,
  median_read_depth = 50000,
  verbose = TRUE
)
}
\arguments{
\item{template}{can be 1) a character string (\code{"Stool"}, \code{"Vaginal"},
or \code{"IBD"}) indicating one of the pre-trained templates in SparseDOSSA2,
or 2) user-provided, fitted results. In the latter case this should be an output
from \code{fit_SparseDOSSA2} or \code{fitCV_SparseDOSSA2}.}

\item{n_sample}{number of samples to simulate}

\item{new_features}{\code{TRUE}/\code{FALSE} indicator for whether or not new
features should be simulated. If \code{FALSE} then the same set of features 
in \code{template} will be simulated.}

\item{n_feature}{number of features to simulate. Only relevant when 
\code{new_features} is \code{TRUE}}

\item{spike_metadata}{for metadata spike-in configurations. Must be one of two things:
a) , 
\itemize{
\item a character string of \code{"none"}, \code{"both"} \code{"abundance"}, 
or \code{"prevalence"}, indicating whether or not
association with metadata will be spiked in. For the spiked-in case, it
indicates if features' abundance/prevalence/both characteristics will be associated
with metadata (also see explanations for \code{metadata_effect_size} and 
\code{perc_feature_spiked_metadata})
\item a data.frame for detailed spike-in configurations. This is the more 
advanced approach, where detailed specification for metadata-microbial 
feature associations are provided. Note: if \code{spike_metadata} is provided
as a data.frame, then \code{metadata_matrix} must be provided as well 
(cannot be generated automatically). In this case, \code{spike_metadata} 
must have exactly four columns: \code{metadata_datum}, \code{feature_spiked}, 
\code{associated_property}, and \code{effect_size}. Each row of the data.frame 
configures one specific metadata-microbe association. Specifically:
\itemize{
\item \code{metadata_datum} (integer) indicates the column number for the metadata
variable to be associated with the microbe
\item \code{feature_spiked} (character) indicates the microbe name to be associated
with the metadata variable
\item \code{associated_property} (character, either \code{"abundance"} or 
\code{"prevalence"}), indicating the property of the microbe to be modified.
If you want the microbe to be associated with the metadata variable
in both properties, include two rows in \code{spike_metadata}, one for 
abundance and one for prevalence
\item \code{effect_size} (numeric) indicating the strength of the association. 
This corresponds to log fold change in non-zero abundance for
\code{"abundance"} spike-in, and log odds ratio for \code{"prevalence"} 
spike-in
}
}}

\item{metadata_effect_size}{(for when \code{spike_metadata} is \code{"abundance"},
\code{"prevalence"}, or \code{"both"}) effect size of the spiked-in associations. This is
non-zero log fold change for abundance spike-in, and log odds ratio for prevalence spike-in}

\item{perc_feature_spiked_metadata}{(for when \code{spike_metadata} is \code{"abundance"},
\code{"prevalence"}, or \code{"both"})
percentage of features to be associated with metadata}

\item{metadata_matrix}{the user can provide a metadata matrix to use for spiking-in
of feature abundances. If using default (\code{NULL}) two variables will be generated: 
one continous, and a binary one of balanced cases and controls. Note: if
\code{spike_metadata} is provided as a data.frame, then the user must provide
\code{metadata_matrix} too}

\item{median_read_depth}{targeted median per-sample read depth}

\item{verbose}{whether detailed information should be printed}
}
\value{
a list with the following component:
\describe{
\item{simulated_data}{
feature by sample matrix of simulated microbial count observations
}
\item{simulated_matrices}{
list of all simulated data matrices, including that of null (i.e. not spiked-in) absolute
abundances, spiked-in absolute abundances, and normalized relative abundances
}
\item{params}{
parameters used for simulation. These are provided in \code{template}.
}
\item{spike_metadata}{
list of variables provided or generated for metadata spike-in. This include 
\code{spike_metadata} for the original \code{spike_metadata} parameter provided
by the user, \code{metadata_matrix} for the
metadata (either provided by the user or internally generated), and 
\code{feature_metadata_spike_df}
for detailed specification of which metadata variables were used to spike-in associations
with which features, in what properties at which effect sizes. This is the 
same as \code{spike_metadata} if the latter was provided as a data.frame.
}
}
}
\description{
\code{SparseDOSSA2} generates synthetic microbial abundance observations
from either pre-trained template, or user-provided fitted results from
\code{fit_SparseDOSSA2} or \code{fitCV_SparseDOSSA2}. Additional options
are available for simulating associations between microbial features
and metadata variables.
}
\examples{
## Using one of the pre-trained SparseDOSSA2 templates:
sim <- SparseDOSSA2(template = "stool", n_sample = 200, new_features = FALSE)
## Using user-provided trained SparseDOSSA2 model:
data("Stool_subset")
fitted <- fit_SparseDOSSA(data = Stool_subset)
sim <- SparseDOSSA2(template = fitted, n_sample = 200, new_features = FALSE)
}
\author{
Siyuan Ma, \email{syma.research@gmail.com}
}
