% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Stool_subset}
\alias{Stool_subset}
\title{A subset of the HMP1-II stool samples}
\format{
A matrix with 5 rows (species) and 5 columns (samples)
}
\source{
\url{https://www.hmpdacc.org/hmp/}
}
\usage{
Stool_subset
}
\description{
A dataset containing species-level microbial counts of a subset of
the HMP1-II stool samples. This includes the top 5 most abundant
species and top 5 most deeply sequenced samples.
}
\keyword{datasets}
