% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_helpers.R
\name{subtract_ifmax}
\alias{subtract_ifmax}
\title{Subtract a fixed value from max counts}
\usage{
subtract_ifmax(y_ij, n_i, value)
}
\arguments{
\item{y_ij}{vector of per-feature read count across samples}
}
\value{
modified counts
}
\description{
Subtract a fixed value from max counts
}
