% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_helpers.R
\name{update_mu}
\alias{update_mu}
\title{M step update for mu}
\usage{
update_mu(sigma2, muhat_ij, sigma2hat_ij, w_i)
}
\arguments{
\item{sigma2}{sigma2 parameters}

\item{muhat_ij}{vector of per-feature muhat estimation across samples}

\item{sigma2hat_ij}{vector of per-feature sigma2hat estimation across samples}

\item{w_i}{vector of observation weights}
}
\value{
updated mu estimation
}
\description{
M step update for mu
}
