% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_helpers.R
\name{update_pi}
\alias{update_pi}
\title{M step update for pi}
\usage{
update_pi(w_i)
}
\arguments{
\item{w_i}{vector of observation weights}
}
\value{
updated pi estimation
}
\description{
M step update for pi
}
