% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{update_sigma2}
\alias{update_sigma2}
\title{M step update for sigma2}
\usage{
update_sigma2(muhat_ij, sigma2hat_ij, w_i, sigma2_old, control)
}
\arguments{
\item{muhat_ij}{vector of per-feature muhat estimation across samples}

\item{sigma2hat_ij}{vector of per-feature sigma2hat estimation across samples}

\item{w_i}{vector of observation weights}

\item{sigma2_old}{old sigma2 parameter (used as initial value for optimization)}

\item{control}{list of control parameters for the inner layer optim function}
}
\value{
updated sigma2 estimation
}
\description{
M step update for sigma2
}
