% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseDOSSA2.R
\name{SparseDOSSA2}
\alias{SparseDOSSA2}
\title{Title}
\usage{
SparseDOSSA2(
  template,
  n_sample,
  new_features = TRUE,
  n_feature,
  spike_metadata = c("none", "both", "abundance", "presence"),
  metadata_effect_size = 1,
  perc_feature_spiked_metadata = 0.05,
  metadata_matrix = NULL,
  feature_metadata_spike_df = NULL,
  median_read_depth = 50000,
  control = list()
)
}
\arguments{
\item{template}{}

\item{n_sample}{}

\item{new_features}{}

\item{n_feature}{}

\item{spike_metadata}{}

\item{metadata_effect_size}{}

\item{perc_feature_spiked_metadata}{}

\item{metadata_matrix}{}

\item{feature_metadata_spike_df}{}

\item{median_read_depth}{}

\item{control}{}
}
\value{

}
\description{
Title
}
