% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseDOSSA2.R
\name{SparseDOSSA2}
\alias{SparseDOSSA2}
\title{Sparse Data Observations for Simulating Synthetic Abundance v2}
\usage{
SparseDOSSA2(n_sample = NULL, n_feature = NULL, template = NULL,
  same_features = FALSE, feature_feature_association = FALSE,
  spike_metadata = NULL, spike_strength = NULL, spike_feature = NULL,
  spike_discrete_bidirectional = TRUE, biomass = NULL,
  read_depth = 50000, seed, control = list(em = list(maxiter.outer =
  1000, maxiter.inner = 1000, reltol.outer = 1e-08, factr.inner = 1e+07),
  kd = list(estimator = "Hscv"), copula = list(method = "spearman", random
  = TRUE, R = 50)), verbose = TRUE)
}
\arguments{
\item{n_sample}{number of samples to simulate}

\item{n_feature}{number of features to simulate}

\item{template}{template dataset to learn community structures from}

\item{same_features}{should SparseDOSSA2 generate random feautures, or the same ones as
in the template dataset?}

\item{feature_feature_association}{only valid when same_feature = TRUE. Should SparseDOSSA2 additionally
model the feature-feature association structure in the template dataset?}

\item{read_depth}{target median sample read depths of the simulated microbial counts}

\item{seed}{random seed}

\item{control}{nested list of control parameters for the various model fitting components of SparseDOSSA2}

\item{verbose}{}
}
\description{
Sparse Data Observations for Simulating Synthetic Abundance v2
}
